requirements.txt
asgiref==3.9.1Django==5.2.5django-cors-headers==4.7.0djangorestframework==3.16.1numpy==2.3.2sqlparse==0.5.3tzdata==2025.2# 🔹 ML & Data Sciencepandas==2.2.3scikit-learn==1.5.2imbalanced-learn==0.12.4django-pandas==0.6.7# 🔹 Utilspython-dotenv==1.0.1# 🔹 Database (only if using PostgreSQL)psycopg2-binary==2.9.9# 🔹 Optional (if you do any plotting)matplotlib==3.9.2